refactor: ReviewCard를 합성 컴포넌트로 변경
