+++ src/features/reviews/components/Animation.style.ts
@@ -1,26 +1,4 @@
-import { css } from "@emotion/react";
 import styled from "@emotion/styled";
-
-import { ReviwCardWithImageProps } from "./ReviewCardWithImage";
-
-export const ReviewCardWithImageContainer = styled.article<
-  Pick<ReviwCardWithImageProps, "isBlock">
->`
-  ${({ isBlock }) =>
-    isBlock
-      ? css`
-          margin: 0 -16px;
-          padding: 16px;
-        `
-      : css`
-          padding: 16px 0;
-        `}
-  border-top: 2px solid ${({ theme }) => theme.colors.neutral[10]};
-
-  &:last-child {
-    border-bottom: 2px solid ${({ theme }) => theme.colors.neutral[10]};
-  }
-`;
 
 export const AnimationConatiner = styled.div`
   display: flex;
