+++ src/features/reviews/components/ReviewComent.tsx
@@ -1,0 +1,9 @@
+import { Comment } from "./ReviewComent.style";
+
+interface ReviewTextProps {
+  text: string;
+}
+
+export default function ReviewComent({ text }: ReviewTextProps) {
+  return <Comment>{text}</Comment>;
+}
