--- src/features/reviews/components/ReviewText.tsx
@@ -1,9 +1,0 @@
-import { Text } from "./ReviewText.style";
-
-interface ReviewTextProps {
-  text: string;
-}
-
-export default function ReviewText({ text }: ReviewTextProps) {
-  return <Text>{text}</Text>;
-}
