+++ src/features/reviews/components/UserRating.style.ts
@@ -1,0 +1,13 @@
+import styled from "@emotion/styled";
+
+export const CreatorContainer = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 12px;
+`;
+
+export const Username = styled.span`
+  ${({ theme }) => theme.typo["body-3-r"]}
+  margin-right: 8px;
+`;
